Dockerized Flask App + reorganization

+++ Yggdrasil-CLI/commands.py
@@ -4,6 +4,7 @@ import os
 import time
 import sqlite3
 import yaml
+from functions import *
 
 RED = "\033[1;31m"
 GREEN = "\033[1;92m"
@@ -31,23 +32,7 @@ null_output = [
 
 # Create the agents.db file and table if it's not already created in the Listeners/http/ directory.
 script_dir = os.path.dirname(os.path.abspath(__file__))
-db_path = os.path.join(script_dir, '..', 'Listeners', 'http', 'agents.db')
-if not os.path.exists(db_path):
-    conn = sqlite3.connect(db_path)
-    cur = conn.cursor()
-
-    cur.execute('''CREATE TABLE IF NOT EXISTS agents (
-        uuid TEXT PRIMARY KEY,
-        name TEXT NOT NULL UNIQUE,
-        status TEXT,
-        first_seen TIMESTAMP,
-        last_seen TIMESTAMP,
-        sleep INTEGER,
-        profile TEXT
-        )
-    ''')
-    conn.commit()
-    conn.close()
+db_path = os.path.join(script_dir, '..', 'Listeners', 'data', 'agents.db')
 
 # Connect to redis database. Redis stores Agent commands
 r = redis.Redis(host="127.0.0.1", port=6379, db=0, decode_responses=True)
@@ -56,45 +41,7 @@ url = f"http://127.0.0.1:8000/admin" # change later to proper port/ip/domain nam
 header = {"Content-Type": "application/json"}
 
 # A clean way to list agents and status.
-def print_table():
-    conn = sqlite3.connect(db_path)
-    cur = conn.cursor()
-    sql_select = """SELECT uuid, name, status, profile FROM agents"""
-    cur.execute(sql_select)
-    result = cur.fetchall()
-    conn.close()
 
-    if not result:
-        print(f"{CYAN}No Callbacks Yet!{RESET}")
-        return
-
-    # Print in a nicely formatted table. Unnecessary but visually better
-    INDEX_WIDTH = 7
-    NAME_WIDTH = 25
-    STATUS_WIDTH = 8
-    PROFILE_WIDTH = 15
-    # Add length of new columns here. Future implement: IP column
-    total_width = INDEX_WIDTH + NAME_WIDTH + STATUS_WIDTH + PROFILE_WIDTH + 13
-    print('=' * total_width)
-    print(f'| {"Index":^{INDEX_WIDTH}} | {"Agent Name":^{NAME_WIDTH}} | {"Profile":^{PROFILE_WIDTH}} | {"Status":^{STATUS_WIDTH}} |') # Print column headers and centers them. 
-    print('=' * total_width)
-
-    # print the table
-    for index, row in enumerate(result):
-        status = row[2]
-        index_cell = str(index).center(INDEX_WIDTH)
-        name_cell = row[1].center(NAME_WIDTH)
-        profile_cell = row[3].center(PROFILE_WIDTH)
-        status_cell = row[2].center(STATUS_WIDTH)
-        if status == "DEAD":
-            print(f"| {index_cell} | {CYAN}{name_cell}{RESET} | {CYAN}{profile_cell}{RESET} | {RED}{status_cell}{RESET} |")
-        elif status == "ALIVE":
-            print(f"| {index_cell} | {CYAN}{name_cell}{RESET} | {CYAN}{profile_cell}{RESET} | {GREEN}{status_cell}{RESET} |")
-        else:
-            print(f"| {index_cell} | {CYAN}{name_cell}{RESET} | {CYAN}{profile_cell}{RESET} | {status_cell} |")
-        print('-' * total_width)
-
-    return result
 
 # Sends agent commands to the Gunicorn/flask app HTTP listener at /admin
 def send_cmd(cmd):
@@ -107,6 +54,7 @@ def send_cmd(cmd):
     try:
         json_payload = {"uuid": id, "command": cmd}
         response = requests.post(url, json=json_payload, headers=header)
+        csv_history(os.getenv('PROFILE'), os.getenv('IP'), cmd)
     except:
         print(f"{RED}ERROR: Failed to send. Is HTTP Listener running?{RESET}")
         return
@@ -139,6 +87,8 @@ def agents(*agrs):
         return
     os.environ['UUID'] = result[agent_index][0]
     os.environ['PROFILE'] = result[agent_index][3]
+    os.environ['IP'] = result[agent_index][4]
+
 
 # Grab uuid stored in environmental variable
 def uuid(*bruh):
@@ -170,12 +120,19 @@ def delete(name):
     if name: # Delete by agent name provided
         conn = sqlite3.connect(db_path)
         cur = conn.cursor()
+        cur.execute("SELECT uuid FROM agents WHERE name = ?", (name,))
+        uuid = cur.fetchone()
+        if not uuid:
+            print(f"{RED}ERROR: Invalid agent name {RESET}")
+            return
+
         cur.execute("DELETE FROM agents WHERE name = ?", (name,))
         conn.commit()
         conn.close()
-
+            
         if os.getenv('UUID') and uuid == os.environ['UUID']:
             del os.environ['UUID']
+
     else: # Delete by index
         agent_list = print_table()
 
@@ -295,6 +252,7 @@ def mass(bruh):
         try:
             json_payload = {"uuid": uuid, "command": command}
             response = requests.post(url, json=json_payload, headers=header)
+            csv_history(result[agent_index[i]][3], result[agent_index[i]][4], command)
         except:
             print(f"{RED}ERROR: Failed to send. Is HTTP Listener running?{RESET}")
             return
